Fix loading of available and configured rules
