+++ src/main/java/com/cloudogu/scm/review/workflow/RepositoryEngineConfigResource.java
@@ -150,57 +150,4 @@ public class RepositoryEngineConfigResource {
     }
     return repository;
   }
-
-  @GET
-  @Path("rules")
-  @Produces(WORKFLOW_MEDIA_TYPE)
-  @Operation(
-    summary = "Workflow engine rules",
-    description = "Returns available rules for the workflow engine.",
-    tags = "Workflow Engine",
-    operationId = "review_get_workflow_rules"
-  )
-  @ApiResponse(
-    responseCode = "200",
-    description = "success",
-    content = @Content(
-      mediaType = MediaType.APPLICATION_JSON,
-      schema = @Schema(implementation = HalRepresentation.class)
-    )
-  )
-  @ApiResponse(
-    responseCode = "500",
-    description = "internal server error",
-    content = @Content(
-      mediaType = VndMediaType.ERROR_TYPE,
-      schema = @Schema(implementation = ErrorDto.class)
-    )
-  )
-  public AvailableRulesDto getAvailableRules(@Context UriInfo uriInfo) {
-    final String selfLink = new PullRequestResourceLinks(uriInfo::getBaseUri).workflowEngineConfigLinks().availableRules();
-    return new AvailableRulesDto(
-      new Links.Builder().self(selfLink).build(),
-      availableRules.stream().map(RuleDto::new).collect(Collectors.toList())
-    );
-  }
-
-  @Getter
-  static class AvailableRulesDto extends HalRepresentation {
-    public AvailableRulesDto(Links links, List<RuleDto> rules) {
-      super(links);
-      this.rules = rules;
-    }
-
-    private final List<RuleDto> rules;
-  }
-
-  @Data
-  static class RuleDto {
-    RuleDto(Rule rule) {
-      this.name = AvailableRules.nameOf(rule);
-      this.applicableMultipleTimes = rule.isApplicableMultipleTimes();
-    }
-    private final String name;
-    private final boolean applicableMultipleTimes;
-  }
 }
